requirements.txt
# ML and Data
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
matplotlib>=3.7.0

# MLflow
mlflow>=2.8.0

# API Serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
evidently>=0.4.0
prometheus-client>=0.17.0
gunicorn>=21.2.0

# Additional utilities
python-dotenv>=1.0.0
